+++ registry/vue/ui/data-table/index.ts
@@ -1,3 +1,3 @@
 export { default as DataTable } from "./DataTable.vue"
 export { default as DataTableColumnHeader } from "./DataTableColumnHeader.vue"
-export { ariaSort, dataTableFeatures, resolveUpdater, type DataTableFeatures } from "./utils"
+export { ariaSort, dataTableFeatures, nextSort, resolveUpdater, type DataTableFeatures } from "./utils"