+++ registry/vue/ui/file-dropzone/index.ts
@@ -1,2 +1,2 @@
 export { default as FileDropzone } from "./FileDropzone.vue"
-export { matchesAccept, partitionFiles, type FileRejection } from "./utils"
+export { limitFiles, matchesAccept, partitionFiles, type FileRejection } from "./utils"